Refactor ID types from Integer to Long in DTOs, models, and repositories; add composite key classes for Favorite and PropertyAmenity

+++ src/main/java/com/trainee/backend_project/dto/PropertyCreateDTO.java
@@ -17,6 +17,6 @@ public class PropertyCreateDTO {
     public String contactName;
     public String contactPhone;
     public String googleMapsUrl;
-    public List<Integer> amenityIds;
+    public List<Long> amenityIds;
     public List<String> imageUrls;
 }
