+++ src/main/java/com/trainee/backend_project/dto/PropertyUpdateDTO.java
@@ -17,6 +17,6 @@ public class PropertyUpdateDTO {
     public String contactName;
     public String contactPhone;
     public String googleMapsUrl;
-    public List<Integer> amenityIds;
+    public List<Long> amenityIds;
     public List<String> imageUrls;
 }
