+++ src/main/java/com/trainee/backend_project/repository/AmenityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AmenityRepository extends JpaRepository<Amenity, Integer> {
+public interface AmenityRepository extends JpaRepository<Amenity, Long> {
 }
