+++ src/main/java/com/trainee/backend_project/repository/PropertyImageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropertyImageRepository extends JpaRepository<PropertyImage, Integer> {
+public interface PropertyImageRepository extends JpaRepository<PropertyImage, Long> {
 }
